Fix almost all clang warnings.
Building with clang using nearly all warnings reported by default by the
QtCreator clang mode model. These additional compile flags are used
(assigned to CFLAGS and CXXFLAGS environment variables):

-fsanitize=address -g -O0 -fno-omit-frame-pointer -Weverything
-Wno-c++98-compat -Wno-c++98-compat-pedantic -Wno-unused-macros
-Wno-newline-eof -Wno-exit-time-destructors -Wno-global-constructors
-Wno-gnu-zero-variadic-macro-arguments -Wno-documentation -Wno-shadow
-Wno-switch-enum -Wno-missing-prototypes -Wno-used-but-marked-unused
-Wno-padded -Wno-documentation-unknown-command -Wno-sign-conversion
-Wno-conversion -Wno-shorten-64-to-32 -Wno-covered-switch-default
-Wno-redundant-parens

+++ src/core/model/dirrenamer.cpp
@@ -348,7 +348,7 @@ void DirRenamer::scheduleAction(TaggedFile* taggedFile)
           // starts with a separator and the search starts with the
           // second character.
           int slashPos = newPart.indexOf(QLatin1Char('/'), 1);
-          if (slashPos != -1 && slashPos != (int)newPart.length() - 1) {
+          if (slashPos != -1 && slashPos != newPart.length() - 1) {
             newPart.truncate(slashPos);
             // the new part has multiple directories
             // => create one directory
@@ -370,7 +370,7 @@ void DirRenamer::scheduleAction(TaggedFile* taggedFile)
         if (newDirname.startsWith(parent)) {
           QString newPart(newDirname.mid(parent.length()));
           int slashPos = newPart.indexOf(QLatin1Char('/'));
-          if (slashPos != -1 && slashPos != (int)newPart.length() - 1) {
+          if (slashPos != -1 && slashPos != newPart.length() - 1) {
             newPart.truncate(slashPos);
             // the new part has multiple directories
             // => rename current directory, then create additional
